Refactor items actions
