--- client/src/modules/list/components/InputBar/model/actions.js
@@ -1,32 +1,0 @@
-import { ENDPOINT_URL } from 'common/constants/variables';
-import { postData } from 'common/utils/fetchMethods';
-import { ItemActionTypes } from './actionTypes';
-import { MessageType as NotificationType } from 'common/constants/enums';
-import { createNotificationWithTimeout } from 'modules/notification/model/actions';
-
-const addItemFailure = errorMessage => ({
-  type: ItemActionTypes.ADD_FAILURE,
-  payload: errorMessage
-});
-export const addItemSuccess = (item, listId) => ({
-  type: ItemActionTypes.ADD_SUCCESS,
-  payload: { item, listId }
-});
-const addItemRequest = () => ({
-  type: ItemActionTypes.ADD_REQUEST
-});
-
-export const addItemToList = (item, listId) => dispatch => {
-  dispatch(addItemRequest());
-  return postData(`${ENDPOINT_URL}/lists/add-item`, { item, listId })
-    .then(resp => resp.json())
-    .then(json => dispatch(addItemSuccess(json, listId)))
-    .catch(err => {
-      dispatch(addItemFailure());
-      createNotificationWithTimeout(
-        dispatch,
-        NotificationType.ERROR,
-        err.message || "Oops, we're sorry, adding item failed..."
-      );
-    });
-};
